Remove unnecessary changed from the migration module
