+++ lib/penthouse/migrator.rb
@@ -2,6 +2,29 @@ require "active_record"
 require "active_support/core_ext/module/aliasing"
 
 module Penthouse
+  module Migration
+    def self.included(base)
+      base.class_eval do
+        # Verbose form of alias_method_chain which is now deprecated in ActiveSupport.
+        #
+        # This replaces the original #annouce method with #announce_with_penthouse
+        # but allows calling #annouce by using #announce_without_penthouse.
+        alias_method :announce_without_penthouse, :announce
+        alias_method :announce, :announce_with_penthouse
+      end
+    end
+
+    def announce_with_penthouse(message)
+      announce_without_penthouse("#{message} - #{current_tenant}")
+    end
+
+    def current_tenant
+      "Tenant: #{Penthouse.tenant || "*** global ***"}"
+    end
+  end
+end
+
+module Penthouse
   module Migrator
     def self.included(base)
       base.extend(ClassMethods)
@@ -114,4 +137,5 @@ module Penthouse
   end
 end
 
-ActiveRecord::Migrator.send(:include, Penthouse::Migrator) if ActiveRecord.version.release < Gem::Version.new("5.2.0")
+ActiveRecord::Migration.send(:include, Penthouse::Migration)
+ActiveRecord::Migrator.send(:include, Penthouse::Migrator)
